% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc_output.R, R/output.R
\name{computeOccupancyProbs}
\alias{computeOccupancyProbs}
\title{computeOccupancyProbs}
\usage{
computeOccupancyProbs(fitmodel)

computeOccupancyProbs(fitmodel)
}
\arguments{
\item{fitmodel}{Output from the function runOccPlus}
}
\value{
An array with the quantiles

An array with the quantiles
}
\description{
Computes the quantiles of the occupancy probability

Computes the quantiles of the occupancy probability
}
\details{
Compute the credible interval of the occupancy probability

Compute the credible interval of the occupancy probability
}
\examples{
\dontrun{
computeOccupancyProbs(fitmodel)
}

\dontrun{
computeOccupancyProbs(fitmodel)
}

}
