% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/occPlus-package.R
\docType{package}
\name{occPlus-package}
\alias{occPlus}
\alias{occPlus-package}
\title{occPlus}
\description{
Package documentation.
}
\author{
\strong{Maintainer}: Alex Diana \email{ad23269@essex.ac.uk}

Authors:
\itemize{
  \item Alex Diana \email{ad23269@essex.ac.uk}
  \item Luveen Wadhwani \email{luveen.w@columbia.edu}
}

}
\keyword{internal}
