% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc_output.R, R/output.R
\name{plotCollectionRates}
\alias{plotCollectionRates}
\title{plotCollectionRates}
\usage{
plotCollectionRates(fitmodel, idx_species = NULL)

plotCollectionRates(fitmodel, idx_species = NULL)
}
\arguments{
\item{fitmodel}{Output from the function runOccPlus}

\item{idx_species}{Indexes of the species to be plotted (leave out to plot all the species).}
}
\value{
The credible interval plot

The credible interval plot
}
\description{
Baseline detection rate for each species.

Baseline detection rate for each species.
}
\details{
Plots the 95% credible interval of the baseline detection rates

Plots the 95% credible interval of the baseline detection rates
}
\examples{
\dontrun{
plotSpeciesRates(fitmodel, idx_species = 1:5)
}

\dontrun{
plotSpeciesRates(fitmodel, idx_species = 1:5)
}

}
