% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc_output.R, R/output.R
\name{plotCorrelationMatrix}
\alias{plotCorrelationMatrix}
\title{plotCorrelationMatrix}
\usage{
plotCorrelationMatrix(fitmodel, idx_species = NULL)

plotCorrelationMatrix(fitmodel, idx_species = NULL)
}
\arguments{
\item{fitmodel}{Output from the function runOccPlus}

\item{idx_species}{Indexes of the species to be plotted (leave out to plot all the species).}
}
\value{
A ggplot object

A ggplot object
}
\description{
Plot the correlation matrix

Plot the correlation matrix
}
\details{
Plots the 95% credible interval of density of true positives and false positives

Plots the 95% credible interval of density of true positives and false positives
}
\examples{
\dontrun{
plotCorrelationMatrix(fitmodel)
}

\dontrun{
plotCorrelationMatrix(fitmodel)
}

}
