% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output.R
\name{plotCumulativeSpeciesDetections}
\alias{plotCumulativeSpeciesDetections}
\title{plotCumulativeSpeciesDetections}
\usage{
plotCumulativeSpeciesDetections(
  fitModel,
  M,
  K,
  primer = 0,
  alpha = 0.95,
  byK = T
)
}
\arguments{
\item{fitModel}{Output from the function runOccJSDM}

\item{M}{Maximum number of environmental samples to consider}

\item{K}{Maximum number of technical replicates (PCRs) to consider}

\item{primer}{Index of the primer to use; 0 (default) pools across all primers}

\item{alpha}{Confidence level of the credible interval, default to .95}

\item{byK}{Should K (technical replicates) be on the x-axis (instead of M)?}
}
\value{
A plot with the credible interval of cumulative detections
}
\description{
Cumulative species detection plot
}
\details{
Plot the credible interval of the overall species detections for several
number of PCR, conditional on species presence in the sample
}
\examples{
\dontrun{
plotCumulativeSpeciesDetections(fitModel, M = 3, K = 5)
}

}
