% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc_output.R, R/output.R
\name{plotDetectionRates}
\alias{plotDetectionRates}
\title{plotDetectionRates}
\usage{
plotDetectionRates(fitmodel, idx_species = NULL)

plotDetectionRates(fitmodel, idx_species = NULL)
}
\arguments{
\item{fitmodel}{Output from the function runOccPlus}

\item{idx_species}{Indexes of the species to be plotted (leave out to plot all the species).}
}
\value{
A ggplot object

A ggplot object
}
\description{
True positives rates at the lab stage for each species.

True positives rates at the lab stage for each species.
}
\details{
Plots the 95% credible interval of the true positives at the lab stage

Plots the 95% credible interval of the true positives at the lab stage
}
\examples{
\dontrun{
plotDetectionRates(fitmodel, idx_species = 1:5)
}

\dontrun{
plotDetectionRates(fitmodel, idx_species = 1:5)
}

}
