% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output.R
\name{plotFPDetectionRates}
\alias{plotFPDetectionRates}
\title{plotFPDetectionRates}
\usage{
plotFPDetectionRates(fitmodel, idx_species = NULL)
}
\arguments{
\item{fitmodel}{Output from the function runOccPlus}

\item{idx_species}{Indexes of the species to be plotted (leave out to plot all the species).}
}
\value{
A ggplot object
}
\description{
False positives rates at the lab stage for each species.
}
\details{
Plots the 95% credible interval of the false positives at the lab stage
}
\examples{
\dontrun{
plotFPDetectionRates(fitmodel, idx_species = 1:5)
}

}
