% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc_output.R, R/output.R
\name{plotOccupancyCovariates}
\alias{plotOccupancyCovariates}
\title{plotOccupancyCovariates}
\usage{
plotOccupancyCovariates(fitmodel, covName = NULL, idx_species = NULL)

plotOccupancyCovariates(fitmodel, covName = NULL, idx_species = NULL)
}
\arguments{
\item{fitmodel}{Output from the function runOccPlus}

\item{covName}{Name of the covariate to be plotted (same name as in data$info)}

\item{idx_species}{Indexes of the species to be plotted (leave out to plot all the species).}
}
\value{
A ggplot object

A ggplot object
}
\description{
Occupancy covariate coefficients.

Occupancy covariate coefficients.
}
\details{
Plots the 95% credible interval of the occupancy covariates coefficients

Plots the 95% credible interval of the occupancy covariates coefficients
}
