% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc_output.R, R/output.R
\name{plotOccupancyRates}
\alias{plotOccupancyRates}
\title{plotOccupancyRates}
\usage{
plotOccupancyRates(fitmodel, idx_species = NULL)

plotOccupancyRates(fitmodel, idx_species = NULL)
}
\arguments{
\item{fitmodel}{Output from the function runOccPlus}

\item{idx_species}{Indexes of the species to be plotted (leave out to plot all the species).}
}
\value{
The credible interval plot

The credible interval plot
}
\description{
Baseline occupancy rate for each species.

Baseline occupancy rate for each species.
}
\details{
Plots the 95% credible interval of the baseline occupancy rates

Plots the 95% credible interval of the baseline occupancy rates
}
\examples{
\dontrun{
plotSpeciesRates(fitmodel, idx_species = 1:5)
}

\dontrun{
plotSpeciesRates(fitmodel, idx_species = 1:5)
}

}
