% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output.R
\name{plotSpeciesResponseCurve}
\alias{plotSpeciesResponseCurve}
\title{plotSpeciesResponseCurve}
\usage{
plotSpeciesResponseCurve(
  species_name,
  target_cov,
  beta_mcmc_j,
  list_matrix,
  raw_df,
  n_points = 200
)
}
\arguments{
\item{fitModel}{Output from the function runOccJSDM}

\item{covName}{Name of the covariate to vary (same name as in data$info)}

\item{idx_species}{Indexes of the species to include (leave out for all species)}

\item{n_grid}{Number of grid points spanning the covariate's range (default 40)}

\item{confidence}{Confidence level of the credible interval, default .95}
}
\value{
A ggplot object
}
\description{
Plot predicted species response curve
}
\examples{
\dontrun{
plotSpeciesResponseCurve(fitModel, covName = "X_psi.1")
}

}
