% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc_output.R, R/output.R
\name{plotStage1FPRates}
\alias{plotStage1FPRates}
\title{plotStage1FPRates}
\usage{
plotStage1FPRates(fitmodel, idx_species = NULL)

plotStage1FPRates(fitmodel, idx_species = NULL)
}
\arguments{
\item{fitmodel}{Output from the function runOccPlus}

\item{idx_species}{Indexes of the species to be plotted (leave out to plot all the species).}
}
\value{
A ggplot object

A ggplot object
}
\description{
False positives rates at the field stage for each species.

False positives rates at the field stage for each species.
}
\details{
Plots the 95% credible interval of the false positives rate at the field stage

Plots the 95% credible interval of the false positives rate at the field stage
}
\examples{
\dontrun{
plotStage1FPRates(fitmodel, idx_species = 1:5)
}

\dontrun{
plotStage1FPRates(fitmodel, idx_species = 1:5)
}

}
