% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc_output.R, R/output.R
\name{plotStage2FPRates}
\alias{plotStage2FPRates}
\title{plotStage2FPRates}
\usage{
plotStage2FPRates(fitmodel, idx_species = NULL)

plotStage2FPRates(fitmodel, idx_species = NULL)
}
\arguments{
\item{fitmodel}{Output from the function runOccPlus}

\item{idx_species}{Indexes of the species to be plotted (leave out to plot all the species).}
}
\value{
A ggplot object

A ggplot object
}
\description{
False positives rates at the lab stage for each species.

False positives rates at the lab stage for each species.
}
\details{
Plots the 95% credible interval of the false positives at the lab stage

Plots the 95% credible interval of the false positives at the lab stage
}
\examples{
\dontrun{
plotStage2FPRates(fitmodel, idx_species = 1:5)
}

\dontrun{
plotStage2FPRates(fitmodel, idx_species = 1:5)
}

}
