% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output.R
\name{returnDetectionCovariates}
\alias{returnDetectionCovariates}
\title{returnDetectionCovariates}
\usage{
returnDetectionCovariates(fitmodel)
}
\arguments{
\item{fitmodel}{Output from the function runOccPlus}
}
\value{
A ggplot object
}
\description{
Detection covariate coefficients.
}
\details{
Returns the detection covariates coefficients posterior sample
}
