% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc_output.R, R/output.R
\name{returnOccupancyCovariates}
\alias{returnOccupancyCovariates}
\title{returnOccupancyCovariates}
\usage{
returnOccupancyCovariates(fitmodel)

returnOccupancyCovariates(fitmodel)
}
\arguments{
\item{fitmodel}{Output from the function runOccPlus}
}
\value{
A ggplot object

A ggplot object
}
\description{
Occupancy covariate coefficients.

Occupancy covariate coefficients.
}
\details{
Returns the occupancy covariates coefficients posterior sample

Returns the occupancy covariates coefficients posterior sample
}
