% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output.R
\name{returnOrdinationCovariatesOutput}
\alias{returnOrdinationCovariatesOutput}
\title{returnOrdinationCovariates}
\usage{
returnOrdinationCovariatesOutput(fitmodel)
}
\arguments{
\item{fitmodel}{Output from the function runOccPlus}
}
\value{
A ggplot object
}
\description{
Ordination covariate coefficients.
}
\details{
Returns the 95% credible interval of the ordination covariates coefficients
}
