% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampleresults.R
\docType{data}
\name{sampleresults}
\alias{sampleresults}
\title{Sample Results Data}
\format{
## `sampleresults`
A data frame with rows and columns:
\describe{
  \item{}{Description of variable}
  \item{}{Description of variable}
  \item{}{Description of variable}
}
}
\source{
occPlus project analysis pipeline
}
\usage{
sampleresults
}
\description{
A dataset containing sample results from the occPlus analysis pipeline.
This data includes processed occupational exposure assessment outputs
and related statistical results.
}
\keyword{datasets}
