% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulateData.R
\name{simulateOccJSDMData}
\alias{simulateOccJSDMData}
\title{simulateOccJSDMData}
\usage{
simulateOccJSDMData(list_datasettings, list_params, list_jsdmParams, model)
}
\arguments{
\item{list_datasettings}{List of data dimension settings (n, S, g, M, P, K,
ncov_psi, ncov_theta).}

\item{list_params}{List of detection/amplification parameters (p, q, theta0,
theta_baseline). Optionally also `mu1`, `sigma1`, `mu0`, `sigma0`
controlling the simulated read-count intensities (see Details); if
omitted these default to `mu1 = 5`, `sigma1 = 1`, `mu0 = 1.5`,
`sigma0 = 1`.}

\item{list_jsdmParams}{List of JSDM parameters (gt, d, ds, sigma_b, sigma_bs,
sigma_ts, sigma_h, l_s).}

\item{model}{Supported data formats are "binary","continuous","occupancy"
and "two_stage"}
}
\value{
A list with simulated data and true parameters.
}
\description{
Simulate data for any supported model type
}
\details{
General-purpose simulation function that supports multiple model types.
}
