% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/occJSDM-package.R
\docType{package}
\name{occJSDM-package}
\alias{occJSDM}
\alias{occJSDM-package}
\title{occJSDM: Occupancy and Joint Species Distribution Models for eDNA Metabarcoding Data}
\value{
This man page documents the package as a whole; it has no
  return value of its own. See \code{\link{runOccJSDM}} for the
  function that fits a model and returns results.
}
\description{
Fits occupancy and joint species distribution models (JSDMs) to
environmental DNA (eDNA) metabarcoding read data within a hierarchical
Bayesian framework, explicitly modeling the two-stage detection process
of field sample collection followed by PCR amplification and
sequencing, while optionally accounting for species traits and spatial
autocorrelation via a Gaussian process. The main entry point for
fitting a model is \code{\link{runOccJSDM}}; \code{\link{simulateOccJSDMData}}
simulates data in the format it expects. See Ji et al. (2025)
\doi{10.1111/ele.70302} for methodological details.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://alexdiana.github.io/occJSDM/}
  \item \url{https://github.com/AlexDiana/occJSDM}
  \item Report bugs at \url{https://github.com/AlexDiana/occJSDM/issues}
}

}
\author{
\strong{Maintainer}: Alex Diana \email{Alex.diana92@yahoo.it}

Authors:
\itemize{
  \item Alex Diana \email{Alex.diana92@yahoo.it}
  \item Douglas W. Yu \email{dougwyu@mac.com}
}

}
\keyword{internal}
