% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output.R
\name{plotCovariateEffect}
\alias{plotCovariateEffect}
\title{plotCovariateEffect}
\usage{
plotCovariateEffect(fitModel, covNames, idx_species, confidence = 0.95)
}
\arguments{
\item{fitModel}{Output from the function runOccJSDM}

\item{idx_species}{Indexes of the species to include (leave out for all species)}

\item{confidence}{Numeric scalar between 0 and 1. The width of the Bayesian
credible interval to compute (default is \code{0.95} for a 95\% interval).}

\item{covName}{Character vector. Name(s) of the covariate(s) to evaluate}
}
\description{
Plot predicted species response curve
}
