% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output.R
\name{plotDetectionCovariates}
\alias{plotDetectionCovariates}
\title{plotDetectionCovariates}
\usage{
plotDetectionCovariates(fitmodel, covName = NULL, idx_species = NULL)
}
\arguments{
\item{fitmodel}{Output from the function runOccPlus}

\item{covName}{Name of the covariate to be plotted (same name as in data$info)}

\item{idx_species}{Indexes of the species to be plotted (leave out to plot all the species).}
}
\value{
A ggplot object
}
\description{
Detection covariate coefficients.
}
\details{
Plots the 95% credible interval of the occupancy covariates coefficients
}
