% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc_output.R, R/output.R
\name{plotFPTPStage2Rates}
\alias{plotFPTPStage2Rates}
\title{plotFPTPStage2Rates}
\usage{
plotFPTPStage2Rates(fitmodel, idx_species = NULL, primerName = NULL)

plotFPTPStage2Rates(fitmodel, idx_species = NULL, primerName = NULL)
}
\arguments{
\item{fitmodel}{Output from the function runOccPlus}

\item{idx_species}{Indexes of the species to be plotted (leave out to plot all the species).}

\item{primerName}{Name of the primer to plot.}
}
\value{
A ggplot object

A ggplot object
}
\description{
True and false positives rates at the lab stage for each species.

True and false positives rates at the lab stage for each species.
}
\details{
Plots the 95% credible interval of the true and false positives at the lab stage

Plots the 95% credible interval of the true and false positives at the lab stage
}
\examples{
\dontrun{
plotFPTPStage2Rates(fitmodel, idx_species = 1:5)
}

\dontrun{
plotFPTPStage2Rates(fitmodel, idx_species = 1:5)
}

}
