% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc_output.R, R/output.R
\name{plotOrdinationCovariates}
\alias{plotOrdinationCovariates}
\title{plotOrdinationCovariates}
\usage{
plotOrdinationCovariates(fitmodel, covName = NULL, idx_factors = NULL)

plotOrdinationCovariates(fitmodel, covName = NULL, idx_factors = NULL)
}
\arguments{
\item{fitmodel}{Output from the function runOccPlus}

\item{covName}{Name of the covariate to be plotted (same name as in data$info)}

\item{idx_species}{Indexes of the species to be plotted (leave out to plot all the species).}
}
\value{
A ggplot object

A ggplot object
}
\description{
Ordination covariate coefficients.

Ordination covariate coefficients.
}
\details{
Plots the 95% credible interval of the ordination covariates coefficients

Plots the 95% credible interval of the ordination covariates coefficients
}
