% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output.R
\name{plotReadIntensity}
\alias{plotReadIntensity}
\title{plotReadIntensity}
\usage{
plotReadIntensity(fitmodel)
}
\arguments{
\item{fitmodel}{Output from the function runOccPlus}
}
\value{
A ggplot object
}
\description{
Plot the reads distribution under the true positives and false positives
}
\details{
Plots the 95% credible interval of density of true positives and false positives
}
\examples{
\dontrun{
plotReadIntensity(fitmodel)
}

}
