% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc_output.R, R/output.R
\name{plotSigElementsCorMatrix}
\alias{plotSigElementsCorMatrix}
\title{plotSigElementsCorMatrix}
\usage{
plotSigElementsCorMatrix(fitmodel, idx_species = NULL)

plotSigElementsCorMatrix(fitmodel, idx_species = NULL)
}
\arguments{
\item{fitmodel}{Output from the function runOccPlus}
}
\value{
A ggplot object

A ggplot object
}
\description{
Plot the significant element of the Correlation matrix

Plot the significant element of the Correlation matrix
}
\details{
Plots the 95% credible interval of density of true positives and false positives

Plots the 95% credible interval of density of true positives and false positives
}
\examples{
\dontrun{
plotSigElementsCorMatrix(fitmodel)
}

\dontrun{
plotSigElementsCorMatrix(fitmodel)
}

}
