% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runMCMCOccPlus.R
\name{runMCMCOccPlus}
\alias{runMCMCOccPlus}
\title{runOccPlus}
\usage{
runMCMCOccPlus(
  data,
  d,
  threshold = F,
  occCovariates = c(),
  ordCovariates = c(),
  detCovariates = c(),
  MCMCparams = list(nchain = 2, nburn = 1000, niter = 1000)
)
}
\arguments{
\item{data}{A list with two elements:
\describe{
  \item{info}{A data.frame of N rows (where N is the total number of samples
  analyzed). Further information in the vignette.}
  \item{OTU}{A matrix of dimension (N x S), where S is the number of species,
  with the number of reads of each species in each sample.}
}}

\item{d}{Number of factors.}

\item{threshold}{(Optional) threshold used to truncated the reads to binary data.
Data greater than or equal to the threshold will be considered as detection.
If this is set to 0 (which is the default), occPlus estimates two modes as described
in the paper.}

\item{occCovariates}{vector of the name of the covariates for the occupancy probabilities.
Names should match the column name in data$info.}

\item{ordCovariates}{vector of the name of the covariates for the occupancy probabilities.
Names should match the column name in data$info.}

\item{detCovariates}{vector of the name of the covariates for the detection probabilities.
Names should match the column name in data$info.}
}
\value{
Description of the return value (e.g., a list, data frame, or numeric output).
}
\description{
Run the OccPlus model.
}
\details{
Fit the model described in Diana et al..
}
\examples{
\dontrun{
# Example usage
fitmodel  <- runOccPlus(data,
d = 2,
occCovariates = c("X_psi.1","X_psi.2"),
ordCovariates = c("X_ord.1","X_ord.2"),
detCovariates = c("X_theta"))
}

}
