% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runOccPlus.R
\name{runOccPlus}
\alias{runOccPlus}
\title{runOccPlus}
\usage{
runOccPlus(
  data,
  d,
  occCovariates = c(),
  ordCovariates = c(),
  detCovariates = c()
)
}
\arguments{
\item{data}{A list with two elements: info.}

\item{d}{Number of factors.}

\item{occCovariates}{vector of the name of the covariates for the occupancy probabilities.
Names should match the column name in data$info.}

\item{ordCovariates}{vector of the name of the covariates for the occupancy probabilities.
Names should match the column name in data$info.}

\item{detCovariates}{vector of the name of the covariates for the detection probabilities.
Names should match the column name in data$info.}
}
\value{
Description of the return value (e.g., a list, data frame, or numeric output).
}
\description{
Run the OccPlus model.
}
\details{
Fit the model described in Diana et al..
}
\examples{
\dontrun{
# Example usage
result <- my_function(param1 = 1, param2 = "text")
}

}
\references{
Optional: Citation or URL for related work.
}
\seealso{
\code{\link{related_function}}
}
